requirements.txt
BeautifulSoup==3.2.1
diff-match-patch==20121119
dj-database-url==0.2.2
dj-static==0.0.5
Django==1.7.1
django-easy-pdf==0.1.0
django-extensions==1.4.9
django-import-export==0.2.7
django-model-report==0.2.1
django-toolbelt==0.0.1
Geraldo==0.4.17
graphviz==0.4.2
gunicorn==18.0
html5lib==0.999
Pillow==2.6.0
pisa==3.0.33
psycopg2==2.5.1
pydot==1.0.2
pyparsing==1.5.7
PyPDF2==1.23
reportlab==2.7
six==1.8.0
static==0.4
stevedore==0.12
tablib==0.10.0
virtualenv==1.10.1
virtualenv-clone==0.2.4
virtualenvwrapper==4.1.1
xhtml2pdf==0.0.6
xlwt==0.7.4
